+++ R/residency.R
@@ -691,7 +691,7 @@ Note:
  if(file.exists("../temp_comments.txt")) cat(gsub("\\r", "", readr::read_file("../temp_comments.txt"))) else cat("No comments were included during the analysis.")
 ```
 
-### Biometric graphics
+', ifelse(biometric.fragment == '', '', paste0('### Biometric graphics
 
 Note:
   : The data used in this graphic is the data present in the biometrics.csv file.
@@ -699,7 +699,7 @@ Note:
 <center>
 ', biometric.fragment,'
 </center>
-
+')), '
 
 ### Last seen
 
@@ -906,8 +906,8 @@ img[src*="#diagram"] {
   <a href="#release-sites">Release sites</a>
   <a href="#array-efficiency">Array efficiency</a>
   <a href="#warning-messages">Warnings</a>
-  <a href="#user-comments">Comments</a>
-  <a href="#biometric-graphics">Biometrics</a>
+  <a href="#user-comments">Comments</a>',
+  ifelse(biometric.fragment == '', '', '\n  <a href="#biometric-graphics">Biometrics</a>'),'
   <a href="#last-seen">Last seen</a>
   <a href="#average-time-of-arrival-at-each-array">Arrival days</a>
   <a href="#time-details-for-each-section">Section times</a>