+++ NEWS.md
@@ -23,6 +23,7 @@ Enhancements:
   * Plot sensor data in the reports.
   * Allow intra-array efficiency to be calculated even if no inter-array efficiency could be calculated.
   * Allow the GUI to pop up if the number of lines to display goes over getOptions("max.print").
+  * Hide biometrics' section from the report if no biometric data was found.
   
 
 ## actel 0.1.3